security-package/issues/84: Introduce ReCaptchaAdminUi module
