--- ReCaptcha/view/adminhtml/layout/adminhtml_auth_forgotpassword.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0"?>
-<!--
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
- -->
-<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-      xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
-
-    <update handle="recaptcha" />
-
-    <body>
-        <referenceContainer name="form.additional.info">
-            <block class="Magento\ReCaptcha\Block\Adminhtml\ReCaptcha" name="recaptcha"
-                   template="Magento_ReCaptcha::recaptcha.phtml"
-                   ifconfig="recaptcha/backend/enabled" />
-        </referenceContainer>
-    </body>
-</page>
