+++ ReCaptchaAdminUi/README.md
@@ -1,0 +1,1 @@
+Please refer to: https://github.com/magento/security-package
