+++ ReCaptchaAdminUi/composer.json
@@ -1,0 +1,28 @@
+{
+    "name": "magento/module-re-captcha-admin-ui",
+    "version": "3.0.0",
+    "description": "Google reCaptcha integration for Magento2",
+    "require": {
+        "php": "~7.1.3||~7.2.0||~7.3.0",
+        "magento/framework": "102.0.*",
+        "magento/module-store": "*",
+        "magento/module-config": "*",
+        "magento/module-re-captcha": "*"
+    },
+    "authors": [
+        {
+            "name": "Riccardo Tempesta",
+            "email": "riccardo.tempesta@magespecialist.it"
+        }
+    ],
+    "type": "magento2-module",
+    "license": "OSL-3.0",
+    "autoload": {
+        "files": [
+            "registration.php"
+        ],
+        "psr-4": {
+            "Magento\\ReCaptchaAdminUi\\": ""
+        }
+    }
+}
