+++ ReCaptchaAdminUi/view/adminhtml/layout/recaptcha.xml
@@ -7,9 +7,8 @@
  -->
 <page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
       xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
-
     <head>
         <script src="https://www.google.com/recaptcha/api.js" src_type="url" />
-        <css src="Magento_ReCaptcha::css/recaptcha.css"/>
+        <css src="Magento_ReCaptchaAdminUi::css/recaptcha.css"/>
     </head>
 </page>
