+++ ReCaptchaAdminUi/view/adminhtml/templates/recaptcha.phtml
@@ -3,7 +3,7 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
-/** @var $block Magento\ReCaptcha\Block\Adminhtml\ReCaptcha */
+/** @var $block Magento\ReCaptchaAdminUi\Block\ReCaptcha */
 $class = $block->getSize() === 'invisible' ? 'field-invisible-recaptcha' : 'field-recaptcha';
 ?>
 <?php if ($block->canDisplayCaptcha()) : ?>
