+++ ReCaptchaAdminUi/view/adminhtml/web/css/recaptcha.less
